refactor: 사용하지 않는 api 정리 및 v2 제거
