+++ backend/src/main/java/todoktodok/backend/discussion/application/dto/request/DiscussionRequest.java
@@ -5,8 +5,8 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
 public record DiscussionRequest(
-        @NotNull(message = "기록 id를 입력해주세요")
-        Long noteId,
+        @NotNull(message = "도서 id를 입력해주세요")
+        Long bookId,
 
         @NotBlank(message = "토론방 제목을 입력해주세요")
         @Size(min = 1, max = 50, message = "토론방 제목은 1자 이상, 50자 이하여야 합니다")
