+++ backend/src/main/java/todoktodok/backend/discussion/application/service/command/DiscussionCommandService.java
@@ -7,15 +7,12 @@ import org.springframework.transaction.annotation.Transactional;
 import todoktodok.backend.book.domain.Book;
 import todoktodok.backend.book.domain.repository.BookRepository;
 import todoktodok.backend.discussion.application.dto.request.DiscussionRequest;
-import todoktodok.backend.discussion.application.dto.request.DiscussionRequestV2;
 import todoktodok.backend.discussion.domain.Discussion;
 import todoktodok.backend.discussion.domain.DiscussionReport;
 import todoktodok.backend.discussion.domain.repository.DiscussionReportRepository;
 import todoktodok.backend.discussion.domain.repository.DiscussionRepository;
 import todoktodok.backend.member.domain.Member;
 import todoktodok.backend.member.domain.repository.MemberRepository;
-import todoktodok.backend.note.domain.Note;
-import todoktodok.backend.note.domain.repository.NoteRepository;
 
 @Service
 @Transactional
@@ -24,7 +21,6 @@ public class DiscussionCommandService {
 
     private final DiscussionRepository discussionRepository;
     private final MemberRepository memberRepository;
-    private final NoteRepository noteRepository;
     private final BookRepository bookRepository;
     private final DiscussionReportRepository discussionReportRepository;
 
@@ -33,32 +29,11 @@ public class DiscussionCommandService {
             final DiscussionRequest discussionRequest
     ) {
         final Member member = findMember(memberId);
-        final Note note = findNote(discussionRequest.noteId());
-
-        validateNoteOwner(note, member);
+        final Book book = findBook(discussionRequest.bookId());
 
         final Discussion discussion = Discussion.builder()
                 .title(discussionRequest.discussionTitle())
                 .content(discussionRequest.discussionOpinion())
-                .member(member)
-                .book(note.getBook())
-                .note(note)
-                .build();
-
-        final Discussion savedDiscussion = discussionRepository.save(discussion);
-        return savedDiscussion.getId();
-    }
-
-    public Long createDiscussionV2(
-            final Long memberId,
-            final DiscussionRequestV2 discussionRequestV2
-    ) {
-        final Member member = findMember(memberId);
-        final Book book = findBook(discussionRequestV2.bookId());
-
-        final Discussion discussion = Discussion.builder()
-                .title(discussionRequestV2.discussionTitle())
-                .content(discussionRequestV2.discussionOpinion())
                 .member(member)
                 .book(book)
                 .build();
@@ -90,11 +65,6 @@ public class DiscussionCommandService {
                 .orElseThrow(() -> new NoSuchElementException("해당 회원을 찾을 수 없습니다"));
     }
 
-    private Note findNote(final Long noteId) {
-        return noteRepository.findById(noteId)
-                .orElseThrow(() -> new NoSuchElementException("해당 기록을 찾을 수 없습니다"));
-    }
-
     private Book findBook(final Long bookId) {
         return bookRepository.findById(bookId)
                 .orElseThrow(() -> new NoSuchElementException("해당 도서를 찾을 수 없습니다"));
@@ -120,15 +90,6 @@ public class DiscussionCommandService {
     ) {
         if (discussion.isOwnedBy(member)) {
             throw new IllegalArgumentException("자기 자신의 토론방을 신고할 수 없습니다");
-        }
-    }
-
-    private void validateNoteOwner(
-            final Note note,
-            final Member member
-    ) {
-        if (!note.isOwnedBy(member)) {
-            throw new IllegalArgumentException("해당 기록의 소유자만 토론방을 생성할 수 있습니다");
         }
     }
 }
