+++ backend/src/main/java/todoktodok/backend/discussion/presentation/DiscussionController.java
@@ -14,12 +14,14 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import todoktodok.backend.discussion.application.dto.request.DiscussionRequest;
 import todoktodok.backend.discussion.application.dto.response.DiscussionResponse;
 import todoktodok.backend.discussion.application.service.command.DiscussionCommandService;
 import todoktodok.backend.discussion.application.service.query.DiscussionQueryService;
+import todoktodok.backend.discussion.domain.DiscussionFilterType;
 import todoktodok.backend.global.auth.Auth;
 import todoktodok.backend.global.auth.Role;
 import todoktodok.backend.global.resolver.LoginMember;
@@ -33,7 +35,6 @@ public class DiscussionController {
     private final DiscussionCommandService discussionCommandService;
     private final DiscussionQueryService discussionQueryService;
 
-    @Deprecated
     @Operation(summary = "토론방 생성 API")
     @Auth(value = Role.USER)
     @PostMapping
@@ -46,16 +47,6 @@ public class DiscussionController {
         return ResponseEntity.status(HttpStatus.CREATED)
                 .location(createUri(discussionId))
                 .build();
-    }
-
-    @Operation(summary = "토론방 전체 조회 API")
-    @Auth(Role.USER)
-    @GetMapping
-    public ResponseEntity<List<DiscussionResponse>> getDiscussions(
-            @Parameter(hidden = true) @LoginMember final Long memberId
-    ) {
-        return ResponseEntity.status(HttpStatus.OK)
-                .body(discussionQueryService.getDiscussions(memberId));
     }
 
     @Operation(summary = "토론방 단일 조회 API")
@@ -80,6 +71,18 @@ public class DiscussionController {
 
         return ResponseEntity.status(HttpStatus.CREATED)
                 .build();
+    }
+
+    @Operation(summary = "토론방 필터링 API")
+    @Auth(value = Role.USER)
+    @GetMapping
+    public ResponseEntity<List<DiscussionResponse>> getDiscussionsByKeywordAndType(
+            @Parameter(hidden = true) @LoginMember final Long memberId,
+            @RequestParam(required = false) final String keyword,
+            @RequestParam final DiscussionFilterType type
+    ) {
+        return ResponseEntity.status(HttpStatus.OK)
+                .body(discussionQueryService.getDiscussionsByKeywordAndType(memberId, keyword, type));
     }
 
     private URI createUri(final Long id) {
