--- backend/src/main/java/todoktodok/backend/shelf/domain/Shelf.java
@@ -1,49 +1,0 @@
-package todoktodok.backend.shelf.domain;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.SQLDelete;
-import org.hibernate.annotations.SQLRestriction;
-import todoktodok.backend.book.domain.Book;
-import todoktodok.backend.global.common.TimeStamp;
-import todoktodok.backend.member.domain.Member;
-
-@Getter
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-
-@Entity
-@SQLRestriction("deleted_at is NULL")
-@SQLDelete(sql = "UPDATE shelf SET deleted_at = NOW() WHERE id = ?")
-public class Shelf extends TimeStamp {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(nullable = false)
-    private Member member;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(nullable = false)
-    private Book book;
-
-    @Builder
-    public static Shelf create(
-            final Member member,
-            final Book book
-    ) {
-        return new Shelf(null, member, book);
-    }
-}
