+++ backend/src/test/java/todoktodok/backend/discussion/domain/DiscussionTest.java
@@ -1,14 +1,13 @@
 package todoktodok.backend.discussion.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import todoktodok.backend.book.domain.Book;
 import todoktodok.backend.book.presentation.fixture.BookFixture;
 import todoktodok.backend.member.domain.Member;
 import todoktodok.backend.member.presentation.fixture.MemberFixture;
-import todoktodok.backend.note.domain.Note;
-import todoktodok.backend.note.presentation.fixture.NoteFixture;
 
 class DiscussionTest {
 
@@ -31,13 +30,6 @@ class DiscussionTest {
                 "1234567890123"
         );
 
-        final Note note = NoteFixture.create(
-                "코드는 다른 사람이 읽을 수도 있다는 사실을 항상 염두에 두어야 한다.",
-                "함수명 하나 짓는 데 시간을 들이는 이유가 명확해졌다.",
-                book,
-                member
-        );
-
         // when - then
         assertThatThrownBy(
                 () -> Discussion.builder()
@@ -45,7 +37,6 @@ class DiscussionTest {
                         .content("클린한 코드를 작성해봅시다")
                         .member(member)
                         .book(book)
-                        .note(note)
                         .build()
         ).isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("토론방 제목은 1자 이상, 50자 이하여야 합니다");
@@ -70,13 +61,6 @@ class DiscussionTest {
                 "1234567890123"
         );
 
-        final Note note = NoteFixture.create(
-                "코드는 다른 사람이 읽을 수도 있다는 사실을 항상 염두에 두어야 한다.",
-                "함수명 하나 짓는 데 시간을 들이는 이유가 명확해졌다.",
-                book,
-                member
-        );
-
         // when - then
         assertThatThrownBy(
                 () -> Discussion.builder()
@@ -84,7 +68,6 @@ class DiscussionTest {
                         .content("클린한 코드를 작성해봅시다")
                         .member(member)
                         .book(book)
-                        .note(note)
                         .build()
         ).isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("토론방 제목은 1자 이상, 50자 이하여야 합니다");
@@ -109,13 +92,6 @@ class DiscussionTest {
                 "1234567890123"
         );
 
-        final Note note = NoteFixture.create(
-                "코드는 다른 사람이 읽을 수도 있다는 사실을 항상 염두에 두어야 한다.",
-                "함수명 하나 짓는 데 시간을 들이는 이유가 명확해졌다.",
-                book,
-                member
-        );
-
         // when - then
         assertThatThrownBy(
                 () -> Discussion.builder()
@@ -123,7 +99,6 @@ class DiscussionTest {
                         .content(content)
                         .member(member)
                         .book(book)
-                        .note(note)
                         .build()
         ).isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("토론방 내용은 1자 이상, 2500자 이하여야 합니다");
@@ -148,13 +123,6 @@ class DiscussionTest {
                 "1234567890123"
         );
 
-        final Note note = NoteFixture.create(
-                "코드는 다른 사람이 읽을 수도 있다는 사실을 항상 염두에 두어야 한다.",
-                "함수명 하나 짓는 데 시간을 들이는 이유가 명확해졌다.",
-                book,
-                member
-        );
-
         // when - then
         assertThatThrownBy(
                 () -> Discussion.builder()
@@ -162,7 +130,6 @@ class DiscussionTest {
                         .content(content)
                         .member(member)
                         .book(book)
-                        .note(note)
                         .build()
         ).isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("토론방 내용은 1자 이상, 2500자 이하여야 합니다");
