+++ backend/src/test/java/todoktodok/backend/discussion/domain/fixture/DiscussionFixture.java
@@ -3,7 +3,6 @@ package todoktodok.backend.discussion.domain.fixture;
 import todoktodok.backend.book.domain.Book;
 import todoktodok.backend.discussion.domain.Discussion;
 import todoktodok.backend.member.domain.Member;
-import todoktodok.backend.note.domain.Note;
 
 public class DiscussionFixture {
 
@@ -11,15 +10,13 @@ public class DiscussionFixture {
             final String title,
             final String content,
             final Member member,
-            final Book book,
-            final Note note
+            final Book book
     ) {
         return Discussion.builder()
                 .title(title)
                 .content(content)
                 .member(member)
                 .book(book)
-                .note(note)
                 .build();
     }
 }
