--- backend/src/test/java/todoktodok/backend/shelf/application/service/query/ShelfQueryServiceTest.java
@@ -1,50 +1,0 @@
-package todoktodok.backend.shelf.application.service.query;
-
-import java.util.List;
-import static org.assertj.core.api.Assertions.assertThat;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.transaction.annotation.Transactional;
-import todoktodok.backend.DatabaseInitializer;
-import todoktodok.backend.InitializerTimer;
-import todoktodok.backend.book.application.dto.response.BookResponse;
-
-@ActiveProfiles("test")
-@Transactional
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
-@ContextConfiguration(initializers = InitializerTimer.class)
-public class ShelfQueryServiceTest {
-
-    @Autowired
-    private DatabaseInitializer databaseInitializer;
-
-    @Autowired
-    private ShelfQueryService shelfQueryService;
-
-    @BeforeEach
-    void setUp() {
-        databaseInitializer.clear();
-    }
-
-    @Test
-    @DisplayName("내 서재의 도서 전체를 조회한다")
-    void getMyBooksTest() {
-        // given
-        databaseInitializer.setDefaultUserInfo();
-        databaseInitializer.setDefaultBookInfo();
-        databaseInitializer.setDefaultShelfInfo();
-
-        final Long memberId = 1L;
-
-        // when
-        final List<BookResponse> myBooks = shelfQueryService.getMyBooks(memberId);
-
-        // then
-        assertThat(myBooks).hasSize(1);
-    }
-}
